feat: The place where it is requested was changed to allow the location of the user
