+++ src/services/api.js
@@ -62,9 +62,7 @@ export const axiosCasesUser = async (id, filterAds) => {
 export const axiosUpdateUser = async (update, id) => {
   try {
     const { data } = await axios.put(`${apiUrl}users/update/${id}`, update);
-
-    document.cookie = "token=" + data;
-
+    localStorage.setItem("token", data);
     return { message: data };
   } catch (error) {
     throw new Error({ message: "The user was not updated" });
